Calendar class extraction implementation
