+++ composer.json
@@ -59,6 +59,7 @@
         "sylius-labs/polyfill-symfony-event-dispatcher": "^1.0.1",
         "sylius-labs/polyfill-symfony-framework-bundle": "^1.0",
         "sylius-labs/polyfill-symfony-security": "^1.0",
+        "sylius/calendar": "v0.1.0",
         "sylius/fixtures-bundle": "^1.6.1",
         "sylius/grid": "^1.8",
         "sylius/grid-bundle": "^1.8",
