+++ src/Sylius/Bundle/ApiBundle/composer.json
@@ -27,6 +27,7 @@
         "doctrine/dbal": "^2.7",
         "api-platform/core": "^2.6",
         "lexik/jwt-authentication-bundle": "^2.6",
+        "sylius/calendar": "v0.1.0",
         "sylius/core-bundle": "^1.7",
         "symfony/messenger": "^4.4 || ^5.4"
     },
