+++ src/Sylius/Bundle/PromotionBundle/Resources/config/services/criteria.xml
@@ -19,7 +19,7 @@
             <tag name="sylius.catalog_promotion.criteria" />
         </service>
         <service id="sylius.catalog_promotion.criteria.date_range" class="Sylius\Bundle\PromotionBundle\Criteria\DateRange">
-            <argument type="service" id="Sylius\Component\Promotion\Provider\DateTimeProviderInterface" />
+            <argument type="service" id="Sylius\Calendar\Provider\CalendarInterface" />
             <tag name="sylius.catalog_promotion.criteria" />
         </service>
     </services>
