+++ src/Sylius/Bundle/PromotionBundle/composer.json
@@ -29,6 +29,7 @@
     "require": {
         "php": "^8.0",
         "stof/doctrine-extensions-bundle": "^1.4",
+        "sylius/calendar": "v0.1.0",
         "sylius/money-bundle": "^1.6",
         "sylius/promotion": "^1.6",
         "sylius/registry": "^1.5",
