--- src/Sylius/Component/Promotion/Provider/DateTimeProviderInterface.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Sylius\Component\Promotion\Provider;
-
-interface DateTimeProviderInterface
-{
-    public function now(): \DateTimeInterface;
-}
